Fix for manual reset problem
Having to press a button or remove a jumper to use the bootloader is a
major hassle. This commit adds support in the BSL script for a
workaround using otherwise unused pins common on FTDI chips: DTR and
RTS. These are both outputs, and since they are not being used for flow
control they can be used to toggle the bootloader pin and the !RESET
line to enable the bootloader.

To have the BSL script automatically invoke the bootloader, pass the -i
flag when calling the script.