Add delay after reset when entering bootloader
On certain Mac platforms, toggling UART pins to enter the bootloader
were happening too fast, causing the CC2538 to not detect that it should
enter the bootloader and ending with the script not working. Adding this
delay takes care of that problem while not affecting the bsl script on
Linux.