add --disable-bootloader option
This lets the script disable the bootloader as the last thing it does.

Adding this option is useful for two reasons:

1. It removes the need to compile an application differently to remove
the bootloader.

2. If the application closes the bootloader, the cc2538-bsl script can
no longer access any memory. This then prevents it from doing anything
else, like writing the secondary address. Making the script close the
bootloader allows it to close it last.