Instantiate a correct object with chip auto-detection
Don't fail if chip ID is unrecognised. If the command succeeds in the first place, then we are most likely talking to something that understand the ROM bootloader's protocol. Ergo, a CC2538, CC13xx or CC26xx.

The return values of COMMAND_GET_CHIP_ID are pretty obscurely documented on the CC13xx/CC26xx TRM, whereas the same command returns something very specific for the CC2538. Thus, we instantiate an object of class CC2538 if we read the corresponding chip ID, default to CC13xx/CC26xx otherwise.