Fix TypeError under PY3
Even though length/4 is always an integer, under PY3 range(0,(length/4)) still fails with TypeError: 'float' object cannot be interpreted as an integer.

This happens because, under PY3:

```
>>> type(4/4)
<class 'float'>
```

Changing the division to a bit shift fixes the error and at the same time we know the calculation result is the same.