Improve CC2640R2 detection
The CC2640R2 has a different logic to determine PG revision number, so our current logic prints incorrect PG. Having said all that, Uniflash and CC2640R2 SDK are reporting different revision numbers for the same device. This commit agrees with the SDK since this is what we have source code for.